@@ -1,9 +1,11 @@
 package com.app.herysapps.calculatordialog;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
 import android.view.View;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.app.herysapps.calculatordialoglib.CalculatorDialog;
 
